CAMBIOS

+++ EntiEspais/EntiEspais/Formularis/FormEspaisTodos.cs
@@ -44,7 +44,72 @@ namespace EntiEspais.Formularis
         //Eliminar
         private void buttonEliminar_Click(object sender, EventArgs e)
         {
-            ORM.EspaisORM.eliminarEspai((ESPAIS)dataGridViewEspais.SelectedRows[0].DataBoundItem);
+            if (dataGridViewEspais.RowCount == 0)
+            {
+                MessageBox.Show("SELECCIONA UN ELEMENT");
+            }
+            else
+            {
+                DialogResult dr = MessageBox.Show("Desitja esborrar aquesta instalació?", "ATENCIÓ", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+
+                if (dr == DialogResult.Yes)
+                {
+                    String mensaje = ORM.EspaisORM.eliminarEspai((ESPAIS)dataGridViewEspais.SelectedRows[0].DataBoundItem);
+
+                    if (mensaje != "")
+                    {
+                        MessageBox.Show(mensaje, "ACCIÓ CANCELADA", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        MessageBox.Show("ESBORRAT");
+                    }
+
+                }
+                else
+                {
+                    MessageBox.Show("ACCIÓ CANCELADA");
+                }
+            }
+        }
+
+        //Modificar al doble click
+        private void dataGridViewEspais_CellDoubleClick(object sender, DataGridViewCellEventArgs e)
+        {
+            Classes.ObridorFormulari.FormEspaisTodosModificar((ESPAIS)dataGridViewEspais.SelectedRows[0].DataBoundItem);
+        }
+
+        //Eliminar al supr
+        private void dataGridViewEspais_UserDeletingRow(object sender, DataGridViewRowCancelEventArgs e)
+        {
+            if (dataGridViewEspais.RowCount == 0)
+            {
+                MessageBox.Show("SELECCIONA UN ELEMENT");
+            }
+            else
+            {
+                DialogResult dr = MessageBox.Show("Desitja esborrar aquesta instalació?", "ATENCIÓ", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+
+                if (dr == DialogResult.Yes)
+                {
+                    String mensaje = ORM.EspaisORM.eliminarEspai((ESPAIS)dataGridViewEspais.SelectedRows[0].DataBoundItem);
+
+                    if (mensaje != "")
+                    {
+                        MessageBox.Show(mensaje, "ACCIÓ CANCELADA", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    else
+                    {
+                        MessageBox.Show("ESBORRAT");
+                    }
+
+                }
+                else
+                {
+                    MessageBox.Show("ACCIÓ CANCELADA");
+                    e.Cancel = true;
+                }
+            }
         }
     }
 }
